feat: Add reading from configs
Add initializing classes with config files.
Add Doxygen documentation

+++ Classes/TDGame/source/Cell.cpp
@@ -1,6 +1,6 @@
 #include "../include/Cell.h"
 namespace TowerDefence {
-	Cell::Cell(const Point& topLeft,CellType type, Building* objectPtr, size_t height, size_t width):
+	Cell::Cell(const Point& topLeft,CellType type, Building* objectPtr, float height, float width):
 		topLeft(topLeft),
 		type(type),
 		objectPtr(objectPtr),
